requirements.txt
langgraph
langchain
langchain-openai
langchain-community
tavily-python
faiss-cpu
streamlit
pandas
matplotlib
python-dotenv
duckduckgo-search
pydantic
sentence-transformers
langchain-anthropic
